mythfrontend: rename commandlineparser.* to mythfrontend_…
